lerp fixes/simple utilities support/ get24TimeFormat method for retrieving world time in 24 hours format

+++ src/main/java/ru/aiefu/timeandwind/IDimType.java
@@ -1,9 +1,9 @@
 package ru.aiefu.timeandwind;
 
 public interface IDimType {
-    void setCycleDuration(double day, double night);
-    double getDayDuration();
-    double getNightDuration();
-    double getCycleDuration();
+    void setCycleDuration(long day, long night);
+    long getDayDuration();
+    long getNightDuration();
+    long getCycleDuration();
     float getPrevAngle();
 }
