+++ src/main/java/ru/aiefu/timeandwind/TimeDataStorage.java
@@ -1,6 +1,6 @@
 package ru.aiefu.timeandwind;
 
 public class TimeDataStorage {
-    public double dayDuration = 12000.0D;
-    public double nightDuration = 12000.0D;
+    public long dayDuration = 12000;
+    public long nightDuration = 12000;
 }
